Instructor review
overall this is a solid start to the project. Pay close attention to how you are attaching event listeners, and how you are querying the DOM. Remember that if you use one of the plural query methods (getElement**s**ByClassName, etc) that you need to find the correct node inside that returned array. Great work overall man!